requirements.txt
streamlit
PyPDF2
python-docx
pandas
openpyxl
whisper
gTTS
plotly
matplotlib
langchain
transformers
faiss-cpu
sentence-transformers
email-validator
twilio
accelerate
tf-keras
langchain-huggingface
pycryptodome
langchain_community
ffmpeg
google-cloud-texttospeech
audio-recorder-streamlit
torch
chardet
